[EGD-5102] Fix warning errors in the BSP
Fix warning errors in the BSP module

+++ module-bsp/board/rt1051/bsp/audio/RT1051BluetoothAudio.cpp
@@ -62,7 +62,7 @@ namespace bsp
             if (framesFetched == 0) {
                 break;
             }
-            else if (framesFetched < inst->audioData.data.size()) {
+            else if (size_t(framesFetched) < inst->audioData.data.size()) {
                 std::fill(inst->audioData.data.begin() + framesFetched, inst->audioData.data.end(), 0);
             }
